+++ CHANGELOG.md
@@ -8,11 +8,15 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Fixed
+
+- Removed dupliate CWE/CVE entities.
+
 ## 2.0.3 - 2020-06-26
 
 ### Fixed
 
-- Add entities to job state 🤦🏼‍♂️
+- Add entities to job state. 🤦🏼‍♂️
 
 ## 2.0.2 - 2020-06-24